moved validator/rpc/gateway under shared,  validator/rpc/web is not visiable to the new shared gateway
