+++ shared/gateway/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
         "gateway.go",
         "log.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/validator/rpc/gateway",
-    visibility = ["//validator:__subpackages__"],
+    importpath = "github.com/prysmaticlabs/prysm/shared/gateway",
+    visibility = ["//visibility:public"],
     deps = [
         "//proto/validator/accounts/v2:ethereum_validator_account_gateway_proto",
         "//validator/web:go_default_library",
