+++ shared/gateway/gateway.go
