+++ validator/web/BUILD.bazel
@@ -12,7 +12,10 @@ go_library(
         ":site_data",  # keep
     ],
     importpath = "github.com/prysmaticlabs/prysm/validator/web",
-    visibility = ["//validator:__subpackages__"],
+    visibility = [
+        "//shared/gateway:__pkg__",
+        "//validator:__subpackages__",
+    ],
     deps = [
         "//shared:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
